This update does some fairly extensive rewriting of pllmpoly02.c, taking the large buffers out of heap memory and putting them in stack memory (also, this makes them available to the debugger).
We seem to have finally resolved our issues with crossing code conversion and can generate lots of homflys from diagrams without a hitch.

+++ src/polynomials.h
@@ -43,7 +43,7 @@
 
 typedef struct monomial_struct {
 
-    double coeff;
+    int coeff;
     int l;  /* Power of l in this monomial */
     int m;  /* Power of m in this monomial */
